Create kind/basic
This is the smallest set of features necessary to make type verifications.
