+++ .travis.sh
@@ -2,6 +2,19 @@
 
 RUBY_V=$(ruby -v)
 
+function run_basic_tests {
+  if [ ! -z "$1" ]; then
+    bundle_cmd="bundle _$1_"
+  else
+    bundle_cmd="bundle"
+  fi
+
+  eval "KIND_BASIC=t $bundle_cmd exec rake test TEST='test/kind/{basic/*_test,basic_test}.rb'"
+  eval "KIND_BASIC=t $bundle_cmd exec rake test TEST='test/kind/{{basic/*_test,basic_test},function_test}.rb'"
+  eval "KIND_BASIC=t $bundle_cmd exec rake test TEST='test/kind/{{basic/*_test,basic_test},either/*_test}.rb'"
+  eval "KIND_BASIC=t $bundle_cmd exec rake test TEST='test/kind/{{basic/*_test,basic_test},result/*_test}.rb'"
+}
+
 function run_with_bundler {
   rm Gemfile.lock
 
@@ -20,16 +33,10 @@ function run_with_am_version_and_bundler {
 }
 
 if [[ $RUBY_V =~ "ruby 2.[12]." ]]; then
-  run_with_bundler "$BUNDLER_V1"
-
   run_with_am_version_and_bundler "3.2" "$BUNDLER_V1"
 fi
 
-RUBY_2_2345="ruby 2.[2345]."
-
-if [[ $RUBY_V =~ $RUBY_2_2345 ]]; then
-  run_with_bundler "$BUNDLER_V1"
-
+if [[ $RUBY_V =~ "ruby 2.[2345]." ]]; then
   run_with_am_version_and_bundler "4.0" "$BUNDLER_V1"
   run_with_am_version_and_bundler "4.1" "$BUNDLER_V1"
   run_with_am_version_and_bundler "4.2" "$BUNDLER_V1"
@@ -38,13 +45,20 @@ if [[ $RUBY_V =~ $RUBY_2_2345 ]]; then
   run_with_am_version_and_bundler "5.2" "$BUNDLER_V1"
 fi
 
+if [[ $RUBY_V =~ "ruby 2.[12345]." ]]; then
+  run_basic_tests "$BUNDLER_V1"
+  run_with_bundler "$BUNDLER_V1"
+fi
+
 RUBY_2_567="ruby 2.[567]."
 RUBY_3_x_x="ruby 3.0."
 
 if [[ $RUBY_V =~ $RUBY_2_567 ]] || [[ $RUBY_V =~ $RUBY_3_x_x ]]; then
   gem install bundler -v ">= 2" --no-doc
 
-  run_with_bundler
   run_with_am_version_and_bundler "6.0"
   run_with_am_version_and_bundler "6.1"
+
+  run_basic_tests
+  run_with_bundler
 fi
