+++ lib/kind/basic/error.rb
@@ -2,6 +2,16 @@
 
 module Kind
   class Error < TypeError
+    INVALID_DEFAULT_ARG = 'the default value must be defined as an argument or block'.freeze
+
+    def self.wrong_number_of_args!(given:, expected:)
+      raise ArgumentError, "wrong number of arguments (given #{given}, expected #{expected})"
+    end
+
+    def self.invalid_default_arg!
+      raise ArgumentError, INVALID_DEFAULT_ARG
+    end
+
     def initialize(arg, object = UNDEFINED, label: nil)
       if UNDEFINED == object
         # Will be used when the exception was raised with a message. e.g:
@@ -13,5 +23,7 @@ module Kind
         super("#{label_text}#{object.inspect} expected to be a kind of #{arg}")
       end
     end
+
+    private_constant :INVALID_DEFAULT_ARG
   end
 end
