+++ lib/kind/basic/kind.rb
