+++ lib/kind/basic/undefined.rb
@@ -1,5 +1,17 @@
 # frozen_string_literal: true
+
 module Kind
+  UNDEFINED = Object.new.tap do |undefined|
+    def undefined.inspect
+      @inspect ||= 'UNDEFINED'.freeze
+    end
+
+    undefined.inspect
+    undefined.freeze
+  end
+
+  private_constant :UNDEFINED
+
   Undefined = Object.new.tap do |undefined|
     def undefined.inspect
       @inspect ||= 'Kind::Undefined'.freeze
