+++ lib/kind/core.rb
@@ -1,19 +1,16 @@
 # frozen_string_literal: true
 
-require 'kind/core/utils/kind'
-require 'kind/core/utils/undefined'
-require 'kind/core/utils/args_error'
+require 'set'
+require 'ostruct'
 
-require 'kind/core/error'
+require 'kind/core/empty'
 require 'kind/core/respond_to'
 require 'kind/core/union_type'
 require 'kind/core/nil'
 require 'kind/core/not_nil'
-require 'kind/core/empty'
 require 'kind/core/dig'
 require 'kind/core/try'
 require 'kind/core/presence'
-require 'kind/core/undefined'
 require 'kind/core/maybe'
 require 'kind/core/type_checker'
 require 'kind/core/type_checkers'
