+++ lib/kind/core/maybe/none.rb
@@ -4,7 +4,7 @@ module Kind
   module Maybe
     class None < Result
       def value_or(default = UNDEFINED, &block)
-        ARGS_ERROR.invalid_default! if UNDEFINED == default && !block
+        Error.invalid_default_arg! if UNDEFINED == default && !block
 
         UNDEFINED != default ? default : block.call
       end
