+++ test/kind/basic/undefined_test.rb
