+++ test/kind/basic_test.rb
@@ -1,6 +1,6 @@
 require 'test_helper'
 
-class Kind::UtilityMethodsTest < Minitest::Test
+class Kind::KindTest < Minitest::Test
   class MyArray < Array
   end
 
@@ -136,24 +136,6 @@ class Kind::UtilityMethodsTest < Minitest::Test
       Kind::Error,
       '"1" expected to be a kind of Module/Class'
     ) { Kind.of_module_or_class('1') }
-  end
-
-  def test_Kind_Of
-    kind_symbol1 = Kind::Of(Symbol)
-
-    assert_instance_of(Kind::TypeChecker::Object, kind_symbol1)
-    assert_equal('Symbol', kind_symbol1.name)
-
-    # -
-
-    kind_symbol2 = Kind::Of(Symbol, name: 'MySymbol')
-
-    assert_instance_of(Kind::TypeChecker::Object, kind_symbol2)
-    assert_equal('MySymbol', kind_symbol2.name)
-
-    # -
-
-    refute_same(kind_symbol1, kind_symbol2)
   end
 
   def test_Kind_value
