+++ test/test_helper.rb
@@ -11,18 +11,22 @@ end
 
 $LOAD_PATH.unshift File.expand_path('../lib', __dir__)
 
-require 'kind'
+unless ENV.fetch('KIND_BASIC', '').empty?
+  require 'kind/basic'
+else
+  require 'kind'
 
-ENV.fetch('ACTIVEMODEL_VERSION', '7.0.0').tap do |active_model_version|
-  if active_model_version < '7.0.0'
-    require 'kind/active_model/validation'
-  end
+  ENV.fetch('ACTIVEMODEL_VERSION', '7.0.0').tap do |active_model_version|
+    if active_model_version < '7.0.0'
+      require 'kind/active_model/validation'
+    end
 
-  if (RUBY_VERSION < '2.2.0' || active_model_version < '4.1')
-    require 'minitest/unit'
+    if (RUBY_VERSION < '2.2.0' || active_model_version < '4.1')
+      require 'minitest/unit'
 
-    module Minitest
-      Test = MiniTest::Unit::TestCase
+      module Minitest
+        Test = MiniTest::Unit::TestCase
+      end
     end
   end
 end
